Split IstioTrafficResourcesYamlParser to Gateway and VirtualService parsers.
